Modified M_SDSP_r_st and M_SDSP_reg_stack_ptr to return constexpr
In DSPEmitter.cpp the functions M_SDSP_r_st and M_SDSP_reg_stack_ptr
return a call of the function MDisp, which is defined in x64Emitter.h
as a constexpr. Because the call of this function accessed the r.st
and reg_stack_ptr arrays at the argument "index", the value could
not be determined at both compile and run time (a requirement for
the constexpr keyword). The change uses pointer arithmetic to
circumvent directly accessing these arrays, based upon the logic
in the dolphin-emu/dolphin repository.

Refs:
 - https://github.com/dolphin-emu/dolphin/blob/master/Source/Core/Core/DSP/Jit/x64/DSPEmitter.cpp#L496-L505

Signed-off-by: Patrick Hickey <pshickey@pshickey.dev>